add methods for visualizing latent neurons
Recreations of the figures seen in the paper showcasing the
activation of the latent neurons with respect to the initial
observations are created using a simple 3D plot.

As training has still yet to incorporate the time evolution
RNN, the results visualized after this commit may or may not be
sensical.