overwrite forward function to add RNN support
The time evolution portion of the Copernicus-modified
Scinet requires a more complex forward function than normal.
The previous quick-fix of adding a simple fully-connected
evolver layer has been replaced with the correct (matching
the paper) simple additive bias time-evolver.

In order to facilitate this network, the `first_pass` parameter
is also added, which defines whether or not the input is passed
through the encoder, as it would in the first iteration of the
RNN, or if it is simply passed through the time evolution.