update training and testing functions
The training method required by the RNN functionality of the
time evolution is a bit more involved than the normal Scinet,
with an extra loop for the time evolution observations.

The testing method, besides the normal calculation of loss, also
outputs the latent and answer outputs of the model throughout
training, in order to facilitate visualization of the network
later.