rearrange the initialization of the loss functions
In order to give access to the leading loss function seperately
from the VAE loss function, it must be initialized outside the
initialization of said VAE loss. The `hyperparameters` code is
reset to the same functionality it had before the VAE update,
and it's supplied loss function is initialized to the optional
`leadingLoss` method, which may be used in some edge cases
instead of the VAE loss.