slightly change the plotting function calls
To keep the process more general, the saving of the plots to
file is moved behind an optional `filename` parameter. The plots
are then "shown" in both functions.
I don't have a preference for whether they should be shown
directly or not, but it's important that they remain consistent
throughout all functions.