add a few optional customizations to `plot_latent`
This commit adds a few optional parameters to the `plot_latent`
function, mostly to allow the recreation of the previous plots
used by `copernicus`.

Namely, the abilities to supply custom axis labels, to apply a
custom tick formatting function and to create a surface plot
are added. This commit should be completely optional and
backwards-compatible