attempt to fix latent plot oddness
Firstly, this commit changes copernicus to finally use the
`visualization` plotting functions rather than it's own. These
functions required a slight bug fix at the same time.

Then, in an attempt to fix the odd "jump" or translation of the
latent activations, the activations of the neurons above
sin(φ) + π are "modulated" back by 2π.

This fix is only partially successful, and more investigation is
required into the cause and fix.