finally fix the latent activation plot jumps
The plots of the latent neuron activations have been plagued by
a ridiculous jump along some unquantifiable sinusoidal curve
in the φ_m direction this entire time. This commit aims to
finally get rid of this bug.

By inspecting the original paper work closer, it was seen that
a few steps were required.

First, the data used for the plots (testing) had to be generated
not randomly, but on a uniform space [0, 2π) along both orbits.

Secondly, any jumps in the time series' evolutions, such as an
orbit passing from -π to π had to be removed, and flattened into
a constantly increasing value.

Finally the actual jump is fixed by an unfortunate brute force
modulation of the data, not along the mysterious sinusoidal
curve as previous attempts had been directed, but by checking
the rows of the dataset against their neighbours.

All this, combined with a change of the plotting to an actual
surface plot, produces the (near) exact results as the paper.